namespace missing
When the user follows the guide, they will not be in their inventory namespace and the inventory route cannot be found. To be consistent with the steps deploying inventory-quarkus, we need to also namespace the request to get the inventory-quarkus route.

As a future improvement I'd recommend using oc project instead of adding -n, except for isolated commands, because with each use of -n the user has to modify the command manually.